metrics handler refactored
