+++ internal/handler/metrics/module.go
@@ -1,4 +1,4 @@
-// Copyright 2022 Dimitrij Drus <dadrus@gmx.de>
+// Copyright 2023 Dimitrij Drus <dadrus@gmx.de>
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,8 +18,13 @@ package metrics
 
 import (
 	"context"
+	"fmt"
+	"net/http"
 
+	"github.com/gofiber/adaptor/v2"
 	"github.com/gofiber/fiber/v2"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/rs/zerolog"
 	"go.uber.org/fx"
 
@@ -27,20 +32,22 @@ import (
 )
 
 var Module = fx.Options( // nolint: gochecknoglobals
-	fx.Provide(fx.Annotated{Name: "metrics", Target: newApp}),
-	fx.Invoke(
-		newHandler,
-		registerHooks,
-	),
+	fx.Invoke(registerHooks),
 )
+
+// ErrLoggerFun is an adapter for promhttp Logger to log errors.
+type ErrLoggerFun func(v ...interface{})
+
+func (l ErrLoggerFun) Println(v ...interface{}) { l(v) }
 
 type hooksArgs struct {
 	fx.In
 
-	Lifecycle fx.Lifecycle
-	Config    *config.Configuration
-	Logger    zerolog.Logger
-	App       *fiber.App `name:"metrics"`
+	Lifecycle  fx.Lifecycle
+	Registerer prometheus.Registerer
+	Gatherer   prometheus.Gatherer
+	Config     *config.Configuration
+	Logger     zerolog.Logger
 }
 
 func registerHooks(args hooksArgs) {
@@ -50,13 +57,31 @@ func registerHooks(args hooksArgs) {
 		return
 	}
 
+	app := fiber.New(fiber.Config{
+		AppName:               "Heimdall's Prometheus endpoint",
+		DisableStartupMessage: true,
+	})
+
+	metricsHandler := promhttp.InstrumentMetricHandler(
+		args.Registerer,
+		promhttp.HandlerFor(
+			args.Gatherer,
+			promhttp.HandlerOpts{
+				Registry: args.Registerer,
+				ErrorLog: ErrLoggerFun(func(v ...interface{}) { args.Logger.Error().Msg(fmt.Sprint(v...)) }),
+			},
+		),
+	)
+
+	registerRoutes(app.Group(args.Config.Metrics.MetricsPath), args.Logger, metricsHandler)
+
 	args.Lifecycle.Append(
 		fx.Hook{
 			OnStart: func(ctx context.Context) error {
 				go func() {
 					addr := args.Config.Metrics.Address()
 					args.Logger.Info().Str("_address", addr).Msg("Metrics service starts listening")
-					if err := args.App.Listen(addr); err != nil {
+					if err := app.Listen(addr); err != nil {
 						args.Logger.Fatal().Err(err).Msg("Could not start Metrics service")
 					}
 				}()
@@ -66,8 +91,14 @@ func registerHooks(args hooksArgs) {
 			OnStop: func(ctx context.Context) error {
 				args.Logger.Info().Msg("Tearing down Metrics service")
 
-				return args.App.Shutdown()
+				return app.Shutdown()
 			},
 		},
 	)
+}
+
+func registerRoutes(router fiber.Router, logger zerolog.Logger, handler http.Handler) {
+	logger.Debug().Msg("Registering Metrics service routes")
+
+	router.Get("", adaptor.HTTPHandler(handler))
 }
